Gmail_Reader_MSVC/requirements.txt
Flask==2.2.3
google_api_python_client==2.80.0
google_auth_oauthlib==1.0.0
pandas==1.5.3
protobuf==4.22.0
requests==2.28.2
